+++ dwds/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add new required parameter `enableDebugging` to `Dwds.start`. If `false` is
   provided, debug services will not run. However, reload logic will continue
   to work with the injected client.
+- Handle injected client SSE errors.
 
 ## 0.5.5